src/test/java/bg/autohouse/validation/offer/TransmissionValidatorTest.java
package bg.autohouse.validation.offer;

import static org.assertj.core.api.Assertions.assertThat;

import org.junit.jupiter.api.Nested;
import org.junit.jupiter.api.Test;

public class TransmissionValidatorTest {
  private TransmissionValidator validator = new TransmissionValidator();

  @Nested
  class ValidFlow {
    @Test
    void isValid_shouldReturnTrue_whenValidTransmission() {
      assertThat(isValid("Manual")).isTrue();
      assertThat(isValid("Automatic")).isTrue();
      assertThat(isValid("SEMI_AUTOMATIC")).isTrue();
    }
  }

  @Nested
  class InvalidFlow {
    @Test
    void isValid_shouldReturnFalse_whenInvalidTransmission() {
      assertThat(isValid("Not valid")).isFalse();
      assertThat(isValid("")).isFalse();
    }
  }

  private boolean isValid(String value) {
    return validator.isValid(value, null);
  }
}
